added options window for geometry and destination folder selection
it needs improving still, the options window doesn't show up the second time the options menu button is clicked, variables are not saved properly.

Improvements have been made to the layout, not more .pack() cause that's ugly, and cumbersome, grid is much more powerful.

The scrollbar has been removed, because who doesn't have a scrollwheel anyways ?

Download/update button is just a dummy for now, it doesn't seem like downloading stuff works, although it may just be that mega is being a pain in the ass (only tested with summertime saga)

config.cfg will be used for all options, and keeping track of which games are installed, and in which version.